Split auto-rule learning into AutoRuleService
Extract auto-rule learning out of CategorizationService into a new
AutoRuleService with a one-way dependency (auto_rule → categorization).
Lazy imports remain only on the bulk_categorize hooks.

- Move record_categorization, confirm, list_pending_proposals,
  list_active_rules, stats, check_overrides, approve, reject into
  AutoRuleService
- Rename methods to match domain language: auto_review →
  list_pending_proposals, auto_confirm → confirm, auto_stats → stats,
  list_auto_rules → list_active_rules
- Add find_matching_rule on CategorizationService as the substrate for
  override detection and future rule-splitting work
- Add app_rule_deactivations audit table and write override events to it
- Wire CLI, MCP tools, and import_service to AutoRuleService
- Cover find_matching_rule and apply_rules provenance with unit tests
- Capture rule-splitting follow-up context in docs/followups.md
