+++ src/moneybin/schema.py
@@ -59,6 +59,7 @@ _SCHEMA_FILES: list[str] = [
     "app_match_decisions.sql",
     "app_seed_source_priority.sql",
     "app_proposed_rules.sql",
+    "app_rule_deactivations.sql",
 ]
