+++ tests/moneybin/test_services/test_auto_rule_service.py
@@ -1,7 +1,6 @@
-"""Unit tests for the auto-rule lifecycle (private _auto_rule module + service surface).
+"""Unit tests for AutoRuleService — proposal lifecycle, override detection, lookups.
 
-Tests intentionally exercise the private ``_auto_rule`` module and
-``CategorizationService._record_categorization`` to assert internal
+Exercises private helpers (``_extract_pattern``) directly to assert internal
 invariants — silencing ``reportPrivateUsage`` for this file is deliberate.
 """
 
@@ -15,8 +14,7 @@ import pytest
 
 from moneybin.config import clear_settings_cache, set_current_profile
 from moneybin.database import Database
-from moneybin.services import _auto_rule
-from moneybin.services.categorization_service import CategorizationService
+from moneybin.services.auto_rule_service import AutoRuleService
 from tests.moneybin.db_helpers import create_core_tables
 
 
@@ -26,7 +24,6 @@ def _mock_db_with_merchant(
     match_type: str = "contains",
 ) -> MagicMock:
     db = MagicMock()
-    # transaction_categories row -> merchant_id
     db.execute.return_value.fetchone.side_effect = [
         (merchant_id,),  # SELECT merchant_id FROM transaction_categories
         (raw_pattern, match_type),  # SELECT raw_pattern, match_type FROM merchants
@@ -37,7 +34,7 @@ def _mock_db_with_merchant(
 def test_extract_pattern_uses_merchant_raw_pattern_when_present() -> None:
     """Extract pattern prefers merchant raw_pattern (matchable substring) when present."""
     db = _mock_db_with_merchant()
-    extracted = _auto_rule.extract_pattern(db, transaction_id="t_1")
+    extracted = AutoRuleService(db)._extract_pattern("t_1")
     assert extracted == ("STARBUCKS", "contains")
 
 
@@ -48,7 +45,7 @@ def test_extract_pattern_falls_back_to_normalized_description() -> None:
         (None,),  # no merchant_id on the categorization row
         ("SQ *STARBUCKS #1234 SEATTLE WA",),  # raw description
     ]
-    extracted = _auto_rule.extract_pattern(db, transaction_id="t_2")
+    extracted = AutoRuleService(db)._extract_pattern("t_2")
     assert extracted == ("STARBUCKS", "contains")
 
 
@@ -56,7 +53,7 @@ def test_extract_pattern_returns_none_when_description_empty() -> None:
     """Extract pattern returns None when description is empty."""
     db = MagicMock()
     db.execute.return_value.fetchone.side_effect = [(None,), ("",)]
-    assert _auto_rule.extract_pattern(db, transaction_id="t_3") is None
+    assert AutoRuleService(db)._extract_pattern("t_3") is None
 
 
 @pytest.fixture
@@ -93,7 +90,7 @@ def _seed_transaction(
 def test_record_creates_proposal_on_first_categorization(real_db: Database) -> None:
     """Creating a proposal on the first categorization stores the expected row."""
     _seed_transaction(real_db, "t1")
-    CategorizationService(real_db)._record_categorization(
+    AutoRuleService(real_db).record_categorization(
         "t1", "Food & Drink", subcategory="Coffee"
     )
 
@@ -109,9 +106,9 @@ def test_record_increments_trigger_count_on_same_pattern_and_category(
     """Repeated categorizations with the same pattern and category increment trigger_count."""
     _seed_transaction(real_db, "t1")
     _seed_transaction(real_db, "t2")
-    svc = CategorizationService(real_db)
-    svc._record_categorization("t1", "Food & Drink", subcategory="Coffee")
-    svc._record_categorization("t2", "Food & Drink", subcategory="Coffee")
+    svc = AutoRuleService(real_db)
+    svc.record_categorization("t1", "Food & Drink", subcategory="Coffee")
+    svc.record_categorization("t2", "Food & Drink", subcategory="Coffee")
 
     rows = real_db.execute(
         "SELECT trigger_count, sample_txn_ids FROM app.proposed_rules"
@@ -127,9 +124,9 @@ def test_record_supersedes_when_same_pattern_different_category(
     """Categorizing a same-pattern txn with a different category supersedes the prior proposal."""
     _seed_transaction(real_db, "t1")
     _seed_transaction(real_db, "t2")
-    svc = CategorizationService(real_db)
-    svc._record_categorization("t1", "Food & Drink")
-    svc._record_categorization("t2", "Groceries")
+    svc = AutoRuleService(real_db)
+    svc.record_categorization("t1", "Food & Drink")
+    svc.record_categorization("t2", "Groceries")
 
     rows = real_db.execute(
         "SELECT category, status FROM app.proposed_rules ORDER BY proposed_at"
@@ -146,7 +143,7 @@ def test_record_skips_when_active_rule_already_covers_pattern(
         "INSERT INTO app.categorization_rules (rule_id, name, merchant_pattern, match_type, category, priority, is_active) "
         "VALUES ('r1', 'starbucks', 'STARBUCKS', 'contains', 'Food & Drink', 100, true)"
     )
-    CategorizationService(real_db)._record_categorization("t1", "Food & Drink")
+    AutoRuleService(real_db).record_categorization("t1", "Food & Drink")
 
     count_row = real_db.execute("SELECT COUNT(*) FROM app.proposed_rules").fetchone()
     assert count_row is not None and count_row[0] == 0
@@ -164,15 +161,15 @@ def test_record_respects_proposal_threshold(
     _seed_transaction(real_db, "t1")
     _seed_transaction(real_db, "t2")
     _seed_transaction(real_db, "t3")
-    svc = CategorizationService(real_db)
-    svc._record_categorization("t1", "Food & Drink")
-    svc._record_categorization("t2", "Food & Drink")
+    svc = AutoRuleService(real_db)
+    svc.record_categorization("t1", "Food & Drink")
+    svc.record_categorization("t2", "Food & Drink")
     pending_row = real_db.execute(
         "SELECT COUNT(*) FROM app.proposed_rules WHERE status = 'pending'"
     ).fetchone()
     assert pending_row is not None and pending_row[0] == 0
 
-    svc._record_categorization("t3", "Food & Drink")
+    svc.record_categorization("t3", "Food & Drink")
     pending_row = real_db.execute(
         "SELECT COUNT(*) FROM app.proposed_rules WHERE status = 'pending'"
     ).fetchone()
@@ -182,11 +179,11 @@ def test_record_respects_proposal_threshold(
 def test_approve_promotes_to_active_rule(real_db: Database) -> None:
     """Approving a pending proposal creates an active rule with the correct attributes."""
     _seed_transaction(real_db, "t1")
-    svc = CategorizationService(real_db)
-    pid = svc._record_categorization("t1", "Food & Drink", subcategory="Coffee")
+    svc = AutoRuleService(real_db)
+    pid = svc.record_categorization("t1", "Food & Drink", subcategory="Coffee")
     assert pid is not None
 
-    result = svc.auto_confirm(approve=[pid])
+    result = svc.confirm(approve=[pid])
     assert result["approved"] == 1
 
     rule = real_db.execute(
@@ -207,14 +204,14 @@ def test_approve_immediately_categorizes_existing_uncategorized(
 ) -> None:
     """Approving a proposal back-fills matching uncategorized transactions immediately."""
     _seed_transaction(real_db, "t1")
-    svc = CategorizationService(real_db)
-    pid = svc._record_categorization("t1", "Food & Drink")
+    svc = AutoRuleService(real_db)
+    pid = svc.record_categorization("t1", "Food & Drink")
     assert pid is not None
     real_db.execute(
         "INSERT INTO core.fct_transactions (transaction_id, account_id, transaction_date, amount, description, source_type) "
         "VALUES ('t9', 'a1', DATE '2026-01-02', -7.00, 'STARBUCKS DOWNTOWN', 'csv')"
     )
-    result = svc.auto_confirm(approve=[pid])
+    result = svc.confirm(approve=[pid])
     assert result["newly_categorized"] == 1
 
     cat = real_db.execute(
@@ -233,10 +230,10 @@ def test_override_threshold_deactivates_rule_and_creates_new_proposal(
 
     # Approve an auto-rule for STARBUCKS -> Food & Drink
     _seed_transaction(real_db, "t1")
-    svc = CategorizationService(real_db)
-    pid = svc._record_categorization("t1", "Food & Drink")
+    svc = AutoRuleService(real_db)
+    pid = svc.record_categorization("t1", "Food & Drink")
     assert pid is not None
-    svc.auto_confirm(approve=[pid])
+    svc.confirm(approve=[pid])
 
     # Two user overrides correcting STARBUCKS to Groceries
     for tid in ("t10", "t11"):
@@ -264,16 +261,22 @@ def test_override_threshold_deactivates_rule_and_creates_new_proposal(
     ).fetchone()
     assert new_proposal == ("Groceries", "pending")
 
+    # Audit row recorded with the override count and the new converged category.
+    audit = real_db.execute(
+        "SELECT reason, override_count, new_category FROM app.rule_deactivations"
+    ).fetchone()
+    assert audit == ("override_threshold", 2, "Groceries")
+
 
 def test_reject_marks_proposal_rejected_without_creating_rule(
     real_db: Database,
 ) -> None:
     """Rejecting a proposal marks it rejected without inserting any categorization rule."""
     _seed_transaction(real_db, "t1")
-    svc = CategorizationService(real_db)
-    pid = svc._record_categorization("t1", "Food & Drink")
+    svc = AutoRuleService(real_db)
+    pid = svc.record_categorization("t1", "Food & Drink")
     assert pid is not None
-    svc.auto_confirm(reject=[pid])
+    svc.confirm(reject=[pid])
 
     status = real_db.execute(
         "SELECT status, decided_by FROM app.proposed_rules WHERE proposed_rule_id = ?",
